requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0

# Redis (optional)
redis>=5.0,<6.0

# Google Gemini AI
google-genai>=0.2,<1.0
Pillow>=10.0,<11.0

# HTTP Client
requests>=2.31,<3.0

# Retry Logic
tenacity>=8.2,<9.0

# Environment Variables
python-dotenv>=1.0,<2.0

# Development
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0

# Static Files (production)
whitenoise>=6.6,<7.0

# Production
gunicorn>=21.2,<22.0
